Fix buffer overflow on empty book name
Empty book name causes "empty" book to be evaluated and everything to be
returned.
That in turn returns more results than the static bound leading to
buffer overflow.

Some test cases include calls to URLs such as:
pismo.cgi?c=%09 ---------------
pismo.cgi?c=%09 .---------------------------------------------
and so on. The parameter can be of almost arbitrary length and contents.